登陆注册新样式,专题页面,文章页面改model,点赞按钮,404页面

+++ mdblog/service/src/main/java/com/mdblog/service/SubjectService.java
@@ -8,7 +8,7 @@ import com.mdblog.po.Subject;
  */
 public interface SubjectService {
     ResponResult getSubject();
-    Subject getSubjectById(Integer SubId);
+    Subject getSubjectById(long SubId);
     // 随机获取一定数量专题
     ResponResult getSubjectRand(Integer num);
 }
